useMongoClient is not necessary!
As I was going thru the tutorials, every time I was running the local server it was giving me an error saying "useMongoClient" is not necessary! And it wanted me to delete it so I think it'll be helpful for the future students to not use it and see all the error that it returns, because it is kind of intimidating at first.